+++ app/layout.tsx
@@ -6,7 +6,6 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html>
       <head>
-        <title>Contentlayer Next.js Example</title>
         <link rel="icon" type="image/x-icon" href="/favicon.png" />
       </head>
       <body>
